fix: 优化布局与目录结构
