+++ Demo/API_V2/Assets/API/Base/GetInfo.meta
