+++ Demo/API_V2/Assets/API/Base/GetInfo/GetInfoSO.asset
@@ -14,20 +14,16 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   entryScriptTypeName: GetInfo
   entryName: "\u83B7\u53D6\u7CFB\u7EDF\u548C\u8BBE\u5907\u4FE1\u606F"
-  entryAPI: "getWindowInfo\uFF5CgetSystemSetting\ngetSystemInfoSync \uFF5CgetSystemInfoAsync\ngetSystemInfo
-    \uFF5CgetDeviceInfo\ngetAppBaseInfo \uFF5CgetAppAuthorizeSetting\ngetEnterOptionsSync
-    \uFF5C getLaunchOptionsSync"
-  entryDescription: 
+  entryAPI: "getSystemSetting\uFF5CgetAppAuthorizeSetting\ngetDeviceInfo\uFF5CgetWindowInfo\ngetAppBaseInfo\ngetEnterOptionsSync\uFF5CgetLaunchOptionsSync"
+  entryDescription: "<color=red>\u4ECE\u57FA\u7840\u5E93 2.20.1 \u5F00\u59CB\uFF0Cwx.getSystemInfo(Sync/Async)\u505C\u6B62\u7EF4\u62A4
+    </color>"
   optionList: []
-  initialButtonText: "\u83B7\u53D6\u7A97\u53E3\u4FE1\u606F"
+  initialButtonText: "\u83B7\u53D6\u8BBE\u5907\u8BBE\u7F6E"
   extraButtonList:
-  - buttonText: "\u83B7\u53D6\u8BBE\u5907\u8BBE\u7F6E  "
-  - buttonText: "\u83B7\u53D6\u7CFB\u7EDF\u4FE1\u606F-\u540C\u6B65"
-  - buttonText: "\u83B7\u53D6\u7CFB\u7EDF\u4FE1\u606F-\u5F02\u6B65"
-  - buttonText: "\u83B7\u53D6\u7CFB\u7EDF\u4FE1\u606F"
-  - buttonText: "\u83B7\u53D6\u8BBE\u5907\u57FA\u7840\u4FE1\u606F"
-  - buttonText: "\u83B7\u53D6\u5FAE\u4FE1APP\u57FA\u7840\u4FE1\u606F"
   - buttonText: "\u83B7\u53D6\u5FAE\u4FE1APP\u6388\u6743\u8BBE\u7F6E"
+  - buttonText: "\u83B7\u53D6\u8BBE\u5907\u57FA\u7840\u4FE1\u606F"
+  - buttonText: "\u83B7\u53D6\u7A97\u53E3\u4FE1\u606F"
+  - buttonText: "\u83B7\u53D6\u5FAE\u4FE1APP\u57FA\u7840\u4FE1\u606F"
   - buttonText: "\u83B7\u53D6\u5C0F\u6E38\u620F\u51B7\u542F\u52A8\u65F6\u7684\u53C2\u6570"
   - buttonText: "\u83B7\u53D6\u5C0F\u6E38\u620F\u6253\u5F00\u7684\u53C2\u6570"
   initialResultList: []
