+++ Demo/API_V2/Assets/API/Network/UDPSocket/UDPSocket.cs
@@ -95,6 +95,7 @@ public class UDPSocket : Details
                 option.message = _bufferData;
             }
             _udpSocket.Write(option);
+            Debug.Log("Message: " + option.message);
         }
         else
         {
