+++ Demo/API_V2/Assets/Scripts/GameManager.cs
@@ -31,6 +31,7 @@ public class GameManager : MonoBehaviour
     // 用于获取微信小游戏的系统信息
     [Header("System Info")]
     public WindowInfo WindowInfo;
+    public ClientRect MenuButtonBoundingClientRect;
 
     private void Awake()
     {
@@ -72,8 +73,10 @@ public class GameManager : MonoBehaviour
                 // 获取微信小游戏的系统信息
 #if UNITY_WEBGL && !UNITY_EDITOR
                 WindowInfo = WX.GetWindowInfo();
+                MenuButtonBoundingClientRect = WX.GetMenuButtonBoundingClientRect();
 #else
                 WindowInfo = new WindowInfo { safeArea = new SafeArea() };
+                MenuButtonBoundingClientRect = new ClientRect();
 #endif
             }
         );
